+++ src/index.ts
@@ -3,6 +3,7 @@ export * from "./components";
 export * from "./hooks";
 export * from "./utils";
 export * from "./types";
+export * from "./api";
 export type {
   MarkerProps,
   PolygonProps,